[FIX] Bug (When Pause is True can't resume cause of the shader)

+++ Assets/Script/BalanceBar.cs
@@ -5,13 +5,13 @@ public class BalanceBar : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxBalance(int balance)
+    public void SetMaxBalance(float balance)
     {
         slider.maxValue = balance;
         slider.value = balance / 2;
     }
 
-    public void SetBalance(int balance)
+    public void SetBalance(float balance)
     {
         slider.value = balance;
     }
